SshHandler introduced
